Misc: use emoji as prefix, no int/long conversions needed

+++ app/src/main/java/com/sduduzog/slimlauncher/data/model/App.kt
@@ -6,7 +6,7 @@ data class App(
         val appName: String,
         val packageName: String,
         val activityName: String,
-        val userSerial : Int
+        val userSerial : Long
 ){
     companion object{
         fun from(homeApp: HomeApp) = App(homeApp.appName, homeApp.packageName, homeApp.activityName, homeApp.userSerial)
